Fix version column type mismatch (INT4 vs INT8) blocking entity loading
The config_versions.version column was INT (INT4) in PostgreSQL but the
Rust code expected i64 (BIGINT/INT8). This caused a decode error that
prevented all entity definitions from loading into the schema cache.

- Change version column to BIGINT in 001_init.sql
- Add migration 004_fix_version_type.sql for existing databases
- Update seed-config.sh to include the new migration

Gateway now successfully loads all 8 entities and their workflows.

+++ migrations/001_init.sql
@@ -94,7 +94,7 @@ CREATE TABLE _atlas.config_versions (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     entity_name VARCHAR(100),
     config_name VARCHAR(100),
-    version INT NOT NULL,
+    version BIGINT NOT NULL,
     config JSONB NOT NULL,
     created_at TIMESTAMPTZ DEFAULT now(),
     created_by UUID REFERENCES _atlas.users(id)
